refactor: manualHandler 관련 코드 제거
