--- app/src/main/java/com/techcourse/ManualHandlerAdapter.java
@@ -1,25 +1,0 @@
-package com.techcourse;
-
-import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerAdapter;
-import com.interface21.webmvc.servlet.view.impl.JspView;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-public class ManualHandlerAdapter implements HandlerAdapter {
-
-    @Override
-    public boolean supports(Object handler) {
-        return handler instanceof ManualHandlerMapping;
-    }
-
-    @Override
-    public ModelAndView handle(Object handler, HttpServletRequest request, HttpServletResponse response)
-            throws Exception {
-        Controller controller = (Controller) ((ManualHandlerMapping) handler).getHandler(request);
-        String viewName = controller.execute(request, response);
-        JspView jspView = new JspView(viewName);
-        return new ModelAndView(jspView);
-    }
-}
